Refactor fixed assets investments to follow new design
