+++ src/main/java/uk/gov/companieshouse/api/accounts/model/entity/smallfull/notes/fixedassetsinvestments/FixedAssetsInvestmentsDataEntity.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.api.accounts.model.entity.notes.fixedassetsinvestments;
+package uk.gov.companieshouse.api.accounts.model.entity.smallfull.notes.fixedassetsinvestments;
 
 import org.springframework.data.mongodb.core.mapping.Field;
 import uk.gov.companieshouse.api.accounts.model.entity.BaseDataEntity;
