+++ src/main/resources/notes.yml
@@ -1,3 +1,4 @@
 notes:
   small-full:
     - off-balance-sheet-arrangements
+    - fixed-assets-investments
